requirements.txt
dj_database_url==0.5.0
Django==4.0.2
django_heroku==0.3.1
django-cors-headers==3.11.0
gunicorn==19.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mysqlclient==2.1.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0